Create 616 - Add Bold Tag in String
Given a string s and a list of strings dict, you need to add a closed pair of bold tag <b> and </b> to wrap the substrings in s that exist in dict. If two such substrings overlap, you need to wrap them together by only one pair of closed bold tag. Also, if two substrings wrapped by bold tags are consecutive, you need to combine them.
Example 1:
Input: 
s = "abcxyz123"
dict = ["abc","123"]
Output:
"<b>abc</b>xyz<b>123</b>"
Example 2:
Input: 
s = "aaabbcc"
dict = ["aaa","aab","bc"]
Output:
"<b>aaabbc</b>c"
Note:
The given dict won't contain duplicates, and its length won't exceed 100.
All the strings in input have length in range [1, 1000].

Approach #3 Using Boolean(Marking) Array
Time complexity : O(l*s*x)O(l∗s∗x). Three nested loops are there to fill boldbold array.
Space complexity : O(s)O(s). resres and boldbold size grows upto O(s)O(s).
Another idea could be to merge the process of identification of the substrings in ss matching with the words in dictdict. To do so, we make use of an array boldboldfor marking the positions of the substrings in ss which are present in dictdict. A True value at bold[i]bold[i] indicates that the current character is a part of the substring which is present in dictdict.
We identify the substrings in ss which are present in dictdict similar to the last approach, by considering only substrings of length length_dlength
​d
​​  for a dictionary word dd. Whenver such a substring is found with its beginning index as ii(and end index (i + length_d -1)(i+length
​d
​​ −1)), we mark all such positions in boldbold as True.
Thus, in this way, whenever a overlapping or consecutive matching substrings exist in ss, a continuous sequence of True values is present in boldbold. Keeping this idea in mind, we traverse over the string ss and keep on putting the current character in the resultant string resres. At every step, we also check if the boldboldarray contains the beginning or end of a continuous sequence of True values. At the beginnning of such a sequence, we put an opening bold tag and then keep on putting the characters of ss till we find a position corresponding to which the last sequence of continuous True values breaks(the first False value is found). We put a closing bold tag at such a position. After this, we again keep on putting the characters of ss in resres till we find the next True value and we keep on continuing the process in the same manner.