Create 351. Android Unlock Patterns Android Unlock Patterns
Given an Android 3x3 key lock screen and two integers m and n, where 1 ≤ m ≤ n ≤ 9, count the total number of unlock patterns of the Android lock screen, which consist of minimum of m keys and maximum n keys.

 

Rules for a valid pattern:

Each pattern must connect at least m keys and at most n keys.
All the keys must be distinct.
If the line connecting two consecutive keys in the pattern passes through any other keys, the other keys must have previously selected in the pattern. No jumps through non selected key is allowed.
The order of keys used matters.
 



 

Explanation:

| 1 | 2 | 3 |
| 4 | 5 | 6 |
| 7 | 8 | 9 |
Invalid move:  4 - 1 - 3 - 6
Line 1-3 passes through key 2 which had not been selected in the pattern.

Invalid move:  4 - 1 - 9 - 2
Line 1-9 passes through key 5 which had not been selected in the pattern.

Valid move:  2 - 4 - 1 - 3 - 6
Line 1-3 is valid because it passes through key 2, which had been selected in the pattern

Valid move:  6 - 5 - 4 - 1 - 9 - 2
Line 1-9 is valid because it passes through key 5, which had been selected in the pattern.

 

Example:

Input: m = 1, n = 1
Output: 9
Credits:
Special thanks to  @elmirap  for adding this problem and creating all test cases.

 

At first glance, this question is so long that it is considered to be a design question, but in fact it is not. This question is still quite interesting, at least closely integrated with the actual situation. This question is about the unlocking method of the Android machine. There are 9 numeric keys. If the length of the password is between [m, n], ask how many unlocking modes there are, and pay attention to the illegal ones given in the question. Sliding mode. So let’s take a look at what is illegal. First, 1 can’t go directly to 3, it must go through 2. Similarly, there are 4 to 6, 7 to 9, 1 to 7, 2 to 8, 3 to 9, and diagonal The line must go through 5, such as 1 to 9, 3 to 7, etc. Create a two-dimensional array jumps to record whether there is an intermediate key between the two numeric keys, and then use a one-bit array visited to record whether a key has been visited, and then use recursion to solve, first call recursion on 1 Function, in a recursive function, traverse each number from 1 to 9 next, and then find if there is a jump number between them. If next has not been accessed and jump is 0, or jump has been accessed, call the recursive function on next. After the number of patterns of number 1 is calculated, since 1, 3, 7, and 9 are symmetrical, you can multiply by 4, and then call the recursive function on number 2. 2, 4, 6, 9 are also symmetrical, and then multiply 4. Finally, call 5 separately, and then add up all of them to get the final result. See the code as follows: