Create 320. Generalized Abbreviation
Write a function to generate the generalized abbreviations of a word. 

Note: The order of the output does not matter.

Example:

Input: "word"
Output:
["word", "1ord", "w1rd", "wo1d", "wor1", "2rd", "w2d", "wo2", "1o1d", "1or1", "w1r1", "1o2", " 2r1", "3d", "w3", "4"]
 

This question allows us to abbreviate part of a word. The rule of abbreviation is that several letters can be represented by numbers, but there can be no two adjacent numbers. For details, please refer to the example given in the title. According to my past experience, This list of all cases must be written in DFS, but I never thought about recursion for a while. Then I counted the number of all cases in the example given in the title, which is 16, and word There are 4 letters, which happen to be 2 to the 4th power. Is this a coincidence? Of course not. Later I found out that if the binary numbers from 0 to 15 are written out, each one can correspond to a situation, as shown below:

0000 word
0001 wor1
0010 wo1d
0011 wo2
0100 w1rd
0101 w1r1
0110 w2d
0111 w3
1000 1ord
1001 1or1
1010 1o1d
1011 1o2
1100 2rd
1101 2r1
1110 3d
1111 4

 

Then we can observe the law, wherever 0 is the original letter, and a single 1 is still 1. If there are several 1s connected together, the number of 1s is required, and this number is used to replace the corresponding letter , Now that the law is found, the code is easy to write, as shown below: