Create 369. Plus One Linked List
Given a non-negative integer represented as non-empty a singly linked list of digits, plus one to the integer.

You may assume the integer do not contain any leading zero, except the number 0 itself.

The digits are stored such that the most significant digit is at the head of the list.

Example:

Input: [1,2,3]
Output: [1,2,4]
 

This question gives us a linked list to simulate a three-digit number. The head of the table is high. Now let us add 1 to the calculation. The difficulty of this question is that the linked list cannot access the elements by coordinates, and can only be traversed. Proceed, and this question just allows us to start from the end of the chain, from back to front, we must deal with it correctly when we encounter a carry, and finally we may have to add a bit at the beginning. Then we think in reverse, if the end of the chain is high, then it is much more convenient to add 1 operation, you can directly traverse and perform arithmetic processing, then what we can do is to flip the linked list first, and now the end of the chain When the high bit is reached, after we add 1 processing operation, we can flip the linked list back, see the code as follows: