Create 261. Graph Valid Tree
Given  n nodes labeled from  0 to  n-1 and a list of undirected edges (each edge is a pair of nodes), write a function to check whether these edges make up a valid tree.

Example 1:

Input:, and 
Output: truen = 5edges = [[0,1], [0,2], [0,3], [1,4]]
Example 2:

Input: and 
Output: falsen = 5, edges = [[0,1], [1,2], [2,3], [1,3], [1,4]]
Note: you can assume that no duplicate edges will appear in  edges. Since all edges are undirected,  [0,1] is the same as  [1,0] and thus will not appear together in  edges.

 

This question gives an undirected graph, let us judge whether it is a tree, if it is a tree, all nodes must be connected, that is to say, it must be a connected graph, and there can be no rings, so the focus is It becomes a verification of whether it is a connected graph and whether it contains a ring. First use DFS to do it, build a graph structure based on pair, represented by adjacency list, and also need a one-bit array v to record whether a node has been visited, and then use DFS to search for node 0, traversed The idea is that when DFS reaches a certain node, first check whether the current node has been visited. If it has been visited, it means that the ring exists, and return false directly. If it has not been visited, mark its status as visited. , And then go to the adjacent link list to find the node adjacent to it to continue recursive traversal. Note that a variable pre is needed to record the previous node so as not to return to the previous node, so that after the traversal is over, the sum node 0 Adjacent nodes are marked as true, and then check whether there are any nodes in v that have not been visited. If there are, it means that the graph is not completely connected, return false, otherwise return true, see the code as follows: