Create 250. Count Univalue Subtrees Count the number of subtrees with the same value
Given a binary tree, count the number of uni-value subtrees.

A Uni-value subtree means all nodes of the subtree have the same value.

Example:

Input: root = [5,1,5,5,5,null,5]

              5
             / \
            1 5
           / \ \
          5 5 5

Output: 4
 

This question allows us to find the number of subtrees with the same value, that is, the number of subtrees with the same value for all nodes. Before, there was a question Largest BST Subtree for finding the largest BST subtree . It feels quite similar , all about subtrees.  For the tree problem, you can also refer to the problem-solving ideas. Recursion can be used here. The first solution method is to traverse all the nodes of the tree in order, and then call for each node to determine the number of words rooted at the current node. To determine whether all nodes are the same, you can refer to the previous question  Same Tree , which uses the idea of ​​divide and conquer, and calls recursion on the left and right words respectively. See the code as follows: