Create 280. Wiggle Sort Wiggle Sort
Given an unsorted array  nums, reorder it in-place such that  nums[0] <= nums[1] >= nums[2] <= nums[3]....

Example:

Input: 
Output: One possible answer is [3,5,1,6,2,4]nums = [3,5,2,1,6,4]
 

This question allows us to find the swing sort. Compared with  Wiggle Sort II  , the conditions for this question are much more relaxed, just because there is an additional equal sign. Due to the existence of the equal sign, it is easy to meet the requirements of the problem when there are repeated numbers in the array. This question first looks at a method with a time complexity of O(nlgn). The idea is to sort the array first, and then only need to change the position of the third number and the second number each time, the fifth number Exchange the position with the fourth number, and so on until the end of the array, so that the swing sort can be completed, see the code as follows: