Create 249. Group Shifted Strings Group Shifted Strings
Given a string, we can "shift" each of its letter to its successive letter, for example:.  "abc" -> "bcd"We can keep "shifting" which forms the sequence:

"abc" -> "bcd" -> ... -> "xyz"
Given a list of strings which contains only lowercase alphabets, group all strings that belong to the same shifting sequence.

Example the For, GIVEN:  ["abc", "bcd", "acef", "xyz", "az", "ba", "a", "z"], 
the Return:

[
  ["abc","bcd","xyz"],
  ["az","ba"],
  ["acef"],
  ["a","z"]
]
Note: For the return value, each  inner  list's elements must follow the lexicographic order.

 
This question lets us reorganize the offset string. The so-called offset string is another string obtained by offsetting each character of a string by the same amount according to the alphabetical order table. The two are offset strings for each other. Note The same string is a special case of offset string, because the offset is 0. Now give a bunch of strings of different lengths, let the offset strings merge together. The blogger first thought of establishing a mapping between the character degree and all offset strings of that length, but it is obvious The error is that the same length is not necessarily offset strings, such as'ab' and'ba', so you can only use HashMap to create a mapping between a string and all sets that are offset strings with this string , Because the title requires the results to be in alphabetical order, so use multiset to save the results, one can save repeated strings, and second, it can be sorted automatically. Then the blogger also wrote a function to determine whether the two strings are offset strings. Note that when comparing the distance between two letters, a trick is used to add 26, and then take the remainder of 26. Traverse the given string set. For the traversed strings, traverse the hash table again, and call the isShifted function to compare with each keyword. If they are offset strings, add the corresponding string set and mark flag, after traversing the HashMap at the end, there is no offset with any keyword, then create a new map, the last thing to do is to convert the multiset to a vector, see the code as follows: