Create 484. Find Permutation
By now, you are given a secret signature consisting of character'D' and'I'.'D' represents a decreasing relationship between two numbers,'I' represents an increasing relationship between two numbers. And our secret signature was constructed by a special integer array, which contains uniquely all the different number from 1 to n (n is the length of the secret signature plus 1). For example, the secret signature "DI" can be constructed by array [2,1,3] or [ 3,1,2], but won't be constructed by array [3,2,4] or [2,1,3,4], which are both illegal constructing special string that can't represent the "DI" secret signature.

On the other hand, now your job is to find the lexicographically smallest permutation of [1, 2, ... n] could refer to the given secret signature in the input.

Example 1:

Input: "I"
Output: [1,2]
Explanation: [1,2] is the only legal initial spectial string can construct secret signature "I", where the number 1 and 2 construct an increasing relationship.
 

Example 2:

Input: "DI"
Output: [2,1,3]
Explanation: Both [2,1,3] and [3,1,2] can construct the secret signature "DI", 
but since we want to find the one with the smallest lexicographical permutation, you need to output [2,1, 3]
 

Note:

The input string will only contain the character'D' and'I'.
The length of input string is a positive integer and will not exceed 10,000
 

This question gave us a string of two characters, D and I, representing the ascending and descending order of the corresponding positions, and we are required to generate the corresponding numeric string based on this string. Due to the influence of the permutation in the name, I feel that the approach should be to find all the permutations and verify them one by one. This method will not pass OJ in all likelihood. In fact, this problem is the easiest to use the greedy algorithm, let's look at an example:

DD II D I

1 2 3 4 5 6 7

3 2 1 4 6 5 7

It is not difficult to see that only the numbers near the position corresponding to D need to be transformed, and the transformation method is to invert the string. What we have to do is to determine the starting position and length of the substring that needs to be inverted through the position of D. can. Through observation, we need to record the starting position i of D, and the continuous number k of D, then we only need to invert the numbers between [i, i+k] in the array, and we can write according to the above ideas code show as below: