Create 266. Palindrome Permutation palindrome
Given a string, determine if a permutation of the string could form a palindrome.

Example 1:

Input: "code"
Output: false
Example 2:

Input: "aab"
Output: true
Example 3:

Input: "carerac"
Output: true
Hint:

Consider the palindromes of odd vs even length. What difference do you notice?
Count the frequency of each character.
If each character occurs even number of times, then it must be a palindrome. How about character which occurs odd number of times?
 

This question allows us to judge whether the entire arrangement of a string is a palindrome. According to the hint in the question, we will discuss the number of strings that are even and even. If it is an even number, due to the palindrome The character of the character string, the number of occurrences of each letter must be an even number of times. When the string is of odd length, only one letter has an odd number of times, and the rest are even numbers. Then we can use this feature to solve the problem, we build The mapping of each letter and its number of occurrences, and then we traverse the HashMap and count the number of letters with odd occurrences. Then there are only two cases where the number of palindromes is the number of palindromes. The first is the letter with no odd occurrences, and the other That is, the length of the string is odd, and there is only one letter with an odd number of times. See the code as follows: