Create 243-Shortest Word Distance
Given a list of words and two words word1 and word2, return the shortest distance between these two words in the list.
For example,
Assume that words = ["practice", "makes", "perfect", "coding", "makes"].
Given word1 = "coding", word2 = "practice", return 3. Given word1 = "makes", word2 = "coding", return 1.
Note:
You may assume that word1 does not equal to word2 , and word1 and word2 are both in the list.
http://segmentfault.com/a/1190000003906667
https://leetcode.com/discuss/50192/java-solution-with- one-for-loop
One pointer points to the last occurrence of word1, and one pointer points to the last occurrence of word2. Because if two words are relatively close, it must be the difference between the positions of adjacent word1 and word2, so we only need to get a new position and compare the position of another word every time.