Create 774-Minimize Max Distance to Gas Station
On a horizontal number line, we have gas stations at positions stations[0], stations[1], ..., stations[N-1], where N = stations.length.

Now, we add K more gas stations so that D, the maximum distance between adjacent gas stations, is minimized.

Return the smallest possible value of D.

Example:

Input: stations = [1, 2, 3 , 4, 5, 6, 7, 8, 9, 10], K = 9
Output: 0.500000
Note:

stations.length will be an integer in range [10, 2000].
stations[i] will be an integer in range [ 0, 10^8].
K will be an integer in range [1, 10^6].
Answers within 10^-6 of the true value will be accepted as correct.

X.
http://www.cnblogs.com/grandyang/p/8970057.html
This question says that we have n gas stations with different distances between the two. Then we can add K gas stations anywhere and ask the minimum value that can make the maximum distance between any two gas stations how many. At first glance, it feels very winding, the biggest for a while, and the smallest for a while. In fact, we can change the scene. For example, n people stand in a team, and the distance between every two people is different. Some people may be very far apart, and some people may be close together. We now need to add K people to the queue again. We hope that the distance between people is as small as possible, so newcomers should join the place where the distance is large, and then ask us after joining K people, ask for the largest distance between people distance. With that said, is it clearer? The blogger first saw the question about the gas station and thought it was related to the previous gas station , but found that there was nothing to do with the two, except that the gas station was shared. For this problem, we still extract the essence, which is the problem of array interpolation. The first thing the blogger considers is to use the greedy algorithm, which is to calculate the distance between every two numbers, and then we insert a number between the two numbers with the largest distance each time. This idea seems correct, but it will Kneeling in such a test case:
[10, 19, 25, 27, 56, 63, 70, 87, 96, 97], K = 3
The distance between the two is:
9,6,2,29,7,7,17,9,1
According to the method mentioned by the blogger, 29 will be separated into two 14.5, and then 17 will be separated into two 8.5s, and there is one gas station left, and one of 14.5 will be separated into two 7.25. But this way, the maximum distance is still 14.5, and in fact, we have a better way. We use two gas stations to divide 29 into three equal parts of 9.67, and then use the remaining one to divide 17. After getting two 8.5s, the maximum distance becomes 9.67. This is the optimal solution. This shows that the greedy algorithm of the blogger's pattern Tucson is not working and lacks respect for Hard topics.
Later, I read the solution in the official answer post and found that the memory used for DP will exceed the standard MLE, and the time used for the heap will exceed the standard TLE. In fact, the correct solution to this question is to use the binary search method. The blogger's first reaction is that there is this kind of operation! ! ? ? This is the operation! ! ! The binary search method used in this question is the fourth in the blogger’s summary post LeetCode Binary Search Summary , that is, the judgment condition of the binary search method is not a simple size relationship, but a situation where sub-functions can be extracted , Let’s see how to get it. If you just want to use the dichotomy to do it, then you must first make it clear what the dichotomy is used to find, is it used to find the location where you want to insert the gas station? Obviously not, it is actually used to find the smallest distance between any two gas stations. This is actually very similar to the previous Kth Smallest Element in a Sorted Matrix . The binary search of that question is to directly halve the number required for positioning, and then verify whether it really meets the meaning of the question. This question is also a similar idea. The range of numbers is given in the question [0, 10^8], then the left and right boundary values ​​of the binary search are available, and the error range is 10^-6, so as long as the difference between right and left If the value is greater than this threshold, the loop continues. The mid we calculated by halving is a candidate, and we need to verify whether a candidate meets the meaning of the question. The method of verification is actually not difficult. We calculate the distance between every two gas stations. If the distance is greater than the candidate, the counter is incremented by 1. If the number greater than the candidate distance is less than or equal to k, then our candidate is too large. , Then right is assigned to mid; conversely, if the number greater than the candidate distance is greater than k, then our candidate is too small, then left is assigned to mid. Finally, left and right will converge to the minimum required maximum distance between any two gas stations. Isn’t it amazing?
http://hehejun.blogspot.com/2018/02/leetcodeminimize-max-distance-to-gas.html
Like Split Array Largest Sum, this problem also has a binary search solution. Because we are sure that the target answer target is in the interval (0, max(stations[i]-stations[i-1])], and given y, we can determine whether it is target in O(n) time Upper/lower bound:
Assuming that the interval is long len, the number of additional gas stations required for the interval to satisfy the maximum value of the subinterval does not exceed y is ceil(len / y), then we only need to calculate the total sum of gas stations required under the condition of y, if sum <= k, which means we can use more gas stations to make the maximum interval smaller, so y is the upper bound of target, target <= y
If sum> k, it means we need to reduce additional gas stations, so y is the upper bound of target, target >= y
Note that the upper and lower bounds are both double types. We set the error range e. When the upper and lower bound error is less than e, we can stop. The time complexity is O(n * log(max(stations[i]-stations[i-1]))), the code is as follows: